feat: Add react-countup and react-intersection-observer dependencies
This commit adds the react-countup and react-intersection-observer dependencies to the project. The react-countup library allows for animated counting of numbers, which will be used to display achievement counts on the landing page. The react-intersection-observer library provides a way to detect when elements are in the viewport, which will be used to trigger the animation of the achievement counts when they become visible to the user. These dependencies enhance the user experience by adding dynamic and interactive elements to the landing page.